+++ src/features/__tests__/ITSystemStatus.test.tsx
@@ -2,11 +2,9 @@ import React from 'react';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { alterMock, render } from 'src/util/test-utils';
-import { ITSystemStatus } from '../employee-only/ITSystemStatus';
+import { ITSystemStatus } from '../it-systems-status/ITSystemStatus';
 import { mockGAEvent } from 'src/setupTests';
 import { IT_STATUS_API } from 'src/mocks/apis';
-
-const mockNoData = { data: [], loading: false, error: false };
 
 describe('<ITSystemStatus /> with issues', () => {
   beforeEach(() => {
