+++ src/features/employee-only/index.ts
@@ -1,4 +1,3 @@
 import { EmployeeTools } from './EmployeeTools';
-import { ITSystemStatus } from './ITSystemStatus';
 
-export { EmployeeTools, ITSystemStatus };
+export { EmployeeTools };
