+++ src/pages/Dashboard/EmployeeDashboard.tsx
@@ -4,7 +4,8 @@ import PageTitle from 'src/ui/PageTitle';
 import EventCardContainer from 'src/ui/EventCardContainer';
 import { MainGridWrapper, Masonry } from 'src/theme';
 import ResourcesCard from 'src/features/ResourcesCard';
-import { EmployeeTools, ITSystemStatus } from 'src/features/employee-only';
+import { EmployeeTools } from 'src/features/employee-only';
+import { ITSystemStatus } from 'src/features/it-systems-status/ITSystemStatus';
 import { FavoriteResources } from 'src/features/FavoriteResources';
 import { TrendingResources } from 'src/features/TrendingResources';
 import { FeaturedTrainingsCard } from 'src/features/training/FeaturedTrainingsCard';
