+++ src/pages/__tests__/Dashboard.test.tsx
@@ -3,21 +3,31 @@ import { screen } from '@testing-library/react';
 import { render, mockEmployeeUser } from 'src/util/test-utils';
 import Dashboard from '../Dashboard';
 
-describe('<Dashboard />', () => {
-  it('renders', async () => {
+describe('<StudentDashboard />', () => {
+  beforeEach(() => {
     render(<Dashboard />);
+  });
+  it('renders the student dashboard', async () => {
     expect(screen.getByTestId('student-dashboard-page')).toBeInTheDocument();
   });
 
-  it('should find "Courses" in the Student Dashboard', async () => {
-    render(<Dashboard />);
-    const courses = await screen.findByText('Courses');
-    expect(courses).toBeInTheDocument();
+  it('should find "Courses" and IT System status card in the Student Dashboard', async () => {
+    expect(await screen.findByText('Courses')).toBeInTheDocument();
+    expect(await screen.findByText(/Major Outage/i)).toBeInTheDocument();
+  });
+});
+
+describe('<EmployeeDashboard />', () => {
+  beforeEach(() => {
+    render(<Dashboard />, { user: mockEmployeeUser });
   });
 
-  it('should find Employee Tools card', async () => {
-    render(<Dashboard />, { user: mockEmployeeUser });
-    const empcenter = await screen.findByText('Employee Tools');
-    expect(empcenter).toBeInTheDocument();
+  it('renders the employee dashboard', async () => {
+    expect(screen.getByTestId('employee-dashboard-page')).toBeInTheDocument();
+  });
+
+  it('should find Employee Tools card and IT System Status', async () => {
+    expect(await screen.findByText('Employee Tools')).toBeInTheDocument();
+    expect(await screen.findByText(/Major Outage/i)).toBeInTheDocument();
   });
 });
